requirements.txt
discord-py-interactions>=5.11.0
python-dotenv>=0.19.1
jurigged
pygit2>=1.13.3
